pyproject.toml
[project]
name = "consulns"
version = "0.1.0"
description = ""
authors = [
  {name = "Luca Tagliavini", email = "luca@teapot.ovh"},
]
license = {file = 'LICENSE'}
readme = "README.md"

dependencies = [
  "click>=8.1.7",
  "pydantic>=2.10.3",
  "pydantic-settings>=2.6.1",
  "rich>=13.9.4",
  "tenacity>=9.0.0",
  "py-consul>=1.5.3",
  "validators>=0.34.0",
  "tabulate>=0.9.0",
  "structlog>=24.4.0",
  "dnspython>=2.7.0"
]
requires-python = ">=3.13"

[project.scripts]
cnsc = "consulns.client:client"
cnsd = "consulns.daemon:daemon"

[project.optional-dependencies]
lint = [
  "ruff>=0.8.0",
  "mypy>=1.13.0",
]
dev = []

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 80
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
select = [
    "F",   # pyflakes
    "E",   # pycodestyle
    "I",   # isort
    "ANN", # flake8 type annotations
    "RUF", # ruff-specific rules
    "FAST" # FastAPI rules
]
fixable = ["ALL"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
disallow_untyped_defs = true
